+++ modules/app-control/src/components/GroupedList.jsx
@@ -21,46 +21,21 @@ const groups = clips =>
     }, {})
   );
 
-class Group extends React.Component {
-  constructor(props) {
-    super(props);
-    this.current = null;
-    this.onPress = this.onPress.bind(this);
-    this.onRelease = this.onRelease.bind(this);
-  }
-
-  onPress(name) {
-    const { onPress, onRelease } = this.props;
-
-    if (this.current) onRelease(this.current);
-    onPress(name);
-    this.current = name;
-  }
-
-  onRelease(name) {
-    this.props.onRelease(name);
-    this.current = null;
-  }
-
-  render() {
-    const { group, pressed } = this.props;
-    return (
-      <div className="PadGroup">
-        {group.clips.map(clip => (
-          <Pad
-            key={clip.id}
-            name={clip.id}
-            pressed={pressed[clip.id]}
-            color={clip.display.color}
-            image={clip.display.cover}
-            onPress={this.onPress}
-            onRelease={this.onRelease}
-          />
-        ))}
-      </div>
-    );
-  }
-}
+const Group = ({ pressed, group, onPress, onRelease }) => (
+  <div className="PadGroup">
+    {group.clips.map(clip => (
+      <Pad
+        key={clip.id}
+        name={clip.id}
+        pressed={pressed[clip.id]}
+        color={clip.display.color}
+        image={clip.display.cover}
+        onPress={onPress}
+        onRelease={onRelease}
+      />
+    ))}
+  </div>
+);
 
 const GroupedList = props => (
   <div className="GroupedList">