+++ modules/audio-playback/src/index.js
@@ -1,6 +1,8 @@
 const Context = require("audio-context");
 const Loader = require("./loader");
 const Player = require("./player");
+const Scheduler = require("./scheduler");
+const Polyphony = require("./polyphony");
 
 const ctx = Context();
 
@@ -9,35 +11,14 @@ function init(set, events) {
 
   const loader = Loader(ctx, events);
   const player = Player(ctx, loader);
-
   loader.load(set.url, clips, set.loader.sources.audio);
 
-  events.on("unmount", function() {
-    console.log("unmount");
-    player.stopAll();
-  });
-
+  Polyphony(set, events, ctx);
+  Scheduler(set, events, player);
   events.on("/audio/query/playing", function() {
     events.emit("/audio/query-results/playing", player.names());
   });
   events.emit("/audio/query-results/playing", []);
-
-  events.on("/clip/stop-all", function() {
-    player.names().forEach(function(name) {
-      events.emit("/clip/stop", name);
-    });
-    events.emit("/audio/all-stopped", name);
-  });
-
-  events.on("/clip/start", function(name) {
-    player.play(name, clips[name], set.audio.defaults);
-    events.emit("/audio/started", name, ctx.currentTime);
-  });
-
-  events.on("/clip/stop", function(name) {
-    player.stop(name);
-    events.emit("/audio/stopped", name, ctx.currentTime);
-  });
 }
 
 module.exports = init;